# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Palmcode CMS API",
        "description": "API documentation for my personal CMS project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
        "/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "operationId": "d534950e818920405745e58047885a95",
                "responses": {
                    "200": {
                        "description": "List of categories returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategorySchema"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a single category by ID",
                "operationId": "785f5317e826dea0958030b8cc5f583d",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategorySchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/v1/pages": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "List all pages",
                "operationId": "b8ed35bb3c9bc32123f2bad00dbbb839",
                "responses": {
                    "200": {
                        "description": "List of pages returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PageSchema"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/pages/{page}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get a single page by ID",
                "operationId": "6e237e883d8033a2138c807933450065",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "ID of the page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page not found"
                    }
                }
            }
        },
        "/v1/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of published posts",
                "operationId": "0599142035606414a3075d42c6cc9441",
                "responses": {
                    "200": {
                        "description": "List of posts returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostSchema"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/posts/{post}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a single post by ID",
                "operationId": "d3297e8db6f632c3170dd6b911e8e627",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CategorySchema": {
                "title": "Category",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Technology"
                    },
                    "slug": {
                        "type": "string",
                        "example": "technology"
                    },
                    "posts_count": {
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-03 16:19:06"
                    }
                },
                "type": "object"
            },
            "PageSchema": {
                "title": "Page",
                "required": [
                    "id",
                    "title"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Contact Us"
                    },
                    "slug": {
                        "type": "string",
                        "example": "contact-us"
                    },
                    "body": {
                        "type": "string",
                        "example": "Content of the page..."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-03 16:19:06"
                    }
                },
                "type": "object"
            },
            "PostSchema": {
                "title": "Post",
                "required": [
                    "id",
                    "title"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Hello World"
                    },
                    "slug": {
                        "type": "string",
                        "example": "hello-world"
                    },
                    "excerpt": {
                        "type": "string",
                        "example": "Short excerpt..."
                    },
                    "status": {
                        "type": "string",
                        "example": "published"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-03T14:48:00.000Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Categories",
            "description": "Categories"
        },
        {
            "name": "Pages",
            "description": "Pages"
        },
        {
            "name": "Posts",
            "description": "Posts"
        }
    ]
}